fix: replace some int() to round() for more precise interpolation of mostly time

+++ src/modules/sbstudio/plugin/operators/takeoff.py
@@ -199,7 +199,7 @@ class TakeoffOperator(StoryboardOperator):
         # Calculate takeoff durations from distances to travel and the
         # average velocity
         fps = context.scene.render.fps
-        takeoff_durations = [ceil((diff / self.velocity) * fps) for diff in diffs]
+        takeoff_durations = [int(ceil((diff / self.velocity) * fps)) for diff in diffs]
 
         # We ensure that drones arrive at the same time, so calculate the
         # takeoff delays for those drones that take off to lower altitudes
